# api-documentation.yaml
openapi: "3.0.3"
info:
  title: "solution API"
  description: "solution API"
  version: "1.0.0"
servers:
  - url: "https://solution"
paths:
...
  /acceptance:
    get:
      summary: "GET acceptance"
      operationId: "listAcceptances"
      parameters:
        - name: "user_id"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AcceptanceDTO"
    post:
      summary: "POST acceptance"
      operationId: "createAcceptance"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptanceDTO"
        required: true
      responses:
        "201":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AcceptanceDTO"
        "404":
          description: "Disclaimer not Found"
components:
  schemas:
...
    AcceptanceDTO:
      type: "object"
      properties:
        disclaimer_id:
          type: "integer"
          format: "int64"
        user_id:
          type: "integer"
          format: "int64"
        create_at:
          type: "integer"
          format: "int64"